+++ NEWS.md
@@ -6,7 +6,9 @@ None
 ## New Features
 1. Fix #61 - Add internal function `hai_data_scale()`
 2. Fix #63 - Add internal function `hai_data_impute()`
-3. Fix #57 - Add internal function `hai_step_trig()`
+3. Fix #58 - Add internal function `hai_step_trig()`
+4. Fix #68 - Add `hai_hyperbolic_vec()` function.
+5. Fix #70 - Add `hai_hyperbolic_augment()` function.
 
 ## Minor Fixes and Improvments
 None